refactor: LottoResult 클래스명을 Ranks로 변경 및 도메인 패키지 구조 변경
