+++ src/main/java/lotto/domain/lotto/LottoTicket.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-import lotto.domain.lotto.utils.Rank;
+import lotto.domain.rank.Rank;
 
 public class LottoTicket {
