+++ src/main/java/lotto/domain/rank/Rank.java
@@ -1,8 +1,9 @@
-package lotto.domain.lotto.utils;
+package lotto.domain.rank;
 
 import java.util.Arrays;
 
 public enum Rank {
+
     FIRST(6, 2_000_000_000),
     SECOND(5, 30_000_000),
     THIRD(5, 1_500_000),
