+++ src/test/java/domain/lotto/RanksTest.java
@@ -3,17 +3,17 @@ package domain.lotto;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.Arrays;
-import lotto.domain.lotto.utils.Rank;
-import lotto.domain.lotto.LottoResult;
+import lotto.domain.rank.Rank;
+import lotto.domain.rank.Ranks;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class LottoResultTest {
+public class RanksTest {
 
     @Test
     @DisplayName("원하는 등수 개수를 제대로 찾아내는지 테스트")
     public void find_wanted_number_of_rank() {
-        LottoResult lottoResult = new LottoResult(
+        Ranks lottoResult = new Ranks(
             Arrays.asList(Rank.FIFTH, Rank.FIFTH, Rank.FIFTH, Rank.SECOND));
         assertThat(lottoResult.getNumberOfRank(Rank.FIRST)).isEqualTo(0);
         assertThat(lottoResult.getNumberOfRank(Rank.SECOND)).isEqualTo(1);
@@ -25,7 +25,7 @@ public class LottoResultTest {
     @Test
     @DisplayName("수익률을 계산한다")
     public void calculate_profit_rate() {
-        LottoResult lottoResult = new LottoResult(
+        Ranks lottoResult = new Ranks(
             Arrays.asList(Rank.FIFTH, Rank.NO_MATCH, Rank.NO_MATCH,
                 Rank.NO_MATCH, Rank.NO_MATCH, Rank.NO_MATCH, Rank.NO_MATCH, Rank.NO_MATCH,
                 Rank.NO_MATCH, Rank.NO_MATCH,
